Remove `GuidWriter`
* Removed `GuidWriter` and moved the writing functionality into the `GenerateGuidRule`.  The additional dependency was not required here, as we pass in the stream writer already
