+++ EndpointForge.Abstractions.Tests/Generators/GuidGeneratorTests.cs
@@ -8,7 +8,7 @@ public class GuidGeneratorTests
     [Fact]
     public void When_New_Expect_Guid_Is_Not_EmptyGuid()
     {
-        var guid = new GuidGenerator().New();
+        var guid = new GuidGenerator().New;
         
         guid.Should().NotBeEmpty();
     }
