+++ EndpointForge.Abstractions/Interfaces/IGuidGenerator.cs
@@ -2,5 +2,5 @@ namespace EndpointForge.Abstractions.Interfaces;
 
 public interface IGuidGenerator
 { 
-    Guid New();
+    Guid New { get; }
 }
