+++ EndpointForge.WebApi.Tests/Fakes/FakeGuidGenerator.cs
@@ -6,5 +6,5 @@ namespace EndpointForge.WebApi.Tests.Fakes;
 [ExcludeFromCodeCoverage]
 internal class FakeGuidGenerator(Guid guid) : IGuidGenerator
 {
-    public Guid New() => guid;
+    public Guid New => guid;
 }
